+++ src/array/codec/array_to_bytes/bytes/bytes_partial_decoder.rs
@@ -25,13 +25,12 @@ impl<'a> BytesPartialDecoder<'a> {
             endian,
         }
     }
-}
 
-impl ArrayPartialDecoderTraits for BytesPartialDecoder<'_> {
-    fn partial_decode(
+    fn do_partial_decode(
         &self,
         decoded_representation: &ArrayRepresentation,
         decoded_regions: &[ArraySubset],
+        parallel: bool,
     ) -> Result<Vec<Vec<u8>>, CodecError> {
         let mut bytes = Vec::with_capacity(decoded_regions.len());
         for array_subset in decoded_regions {
@@ -56,7 +55,11 @@ impl ArrayPartialDecoderTraits for BytesPartialDecoder<'_> {
                 let mut bytes_subset = decoded.concat();
                 if let Some(endian) = &self.endian {
                     if !endian.is_native() {
-                        reverse_endianness(&mut bytes_subset, decoded_representation.data_type());
+                        reverse_endianness(
+                            &mut bytes_subset,
+                            decoded_representation.data_type(),
+                            parallel,
+                        );
                     }
                 }
                 bytes_subset
@@ -70,5 +73,23 @@ impl ArrayPartialDecoderTraits for BytesPartialDecoder<'_> {
             bytes.push(bytes_subset);
         }
         Ok(bytes)
+    }
+}
+
+impl ArrayPartialDecoderTraits for BytesPartialDecoder<'_> {
+    fn partial_decode(
+        &self,
+        decoded_representation: &ArrayRepresentation,
+        decoded_regions: &[ArraySubset],
+    ) -> Result<Vec<Vec<u8>>, CodecError> {
+        self.do_partial_decode(decoded_representation, decoded_regions, false)
+    }
+
+    fn par_partial_decode(
+        &self,
+        decoded_representation: &ArrayRepresentation,
+        decoded_regions: &[ArraySubset],
+    ) -> Result<Vec<Vec<u8>>, CodecError> {
+        self.do_partial_decode(decoded_representation, decoded_regions, true)
     }
 }